Add Vue Router and shift app.vue into home.vue
