+++ frontend/package.json
@@ -30,6 +30,7 @@
     "vue": "^3.2.47",
     "vue-loading-overlay": "^5.0.3",
     "vue-observe-visibility": "^1.0.0",
+    "vue-router": "^4.1.6",
     "vuex": "^4.0.2",
     "zod": "^3.20.6"
   },
